+++ src/algorithms.h
