+++ src/interactive/interactive_mode.h
