+++ src/priority_queues/priority_queue.h
@@ -12,19 +12,19 @@
 template<typename T>
 struct priority_queue {
 protected:
-    comparator<T> sort_comparator;
+	comparator<T> sort_comparator;
 
 public:
-    /*!
-     * Constructs a priority queue with default behaviour.
-     */
-    priority_queue();
-    /*!
-     * Constructs a priority queue with a custom comparator function.
-     * The function will be used to sort priorities of elements.
-     * \param sort_comparator the comparator function to use for sorting priorities
-     */
-    explicit priority_queue(comparator<T>& sort_comparator);
+	/*!
+	 * Constructs a priority queue with default behaviour.
+	 */
+	priority_queue();
+	/*!
+	 * Constructs a priority queue with a custom comparator function.
+	 * The function will be used to sort priorities of elements.
+	 * \param sort_comparator the comparator function to use for sorting priorities
+	 */
+	explicit priority_queue(comparator<T>& sort_comparator);
 	virtual ~priority_queue() = default;
 
 	/*!
@@ -33,16 +33,16 @@ public:
 	 */
 	virtual void add(const T& element) = 0;
 
-    /*!
-     * Remove the highest-priority element and return its value.
-     * \return the element with the highest priority value
-     */
+	/*!
+	 * Remove the highest-priority element and return its value.
+	 * \return the element with the highest priority value
+	 */
 	virtual T pop() = 0;
-    /*!
-     * Return a reference to the highest-priority element. This is an O(1) operation.
-     * Keep in mind that the reference might be invalidated after modifying the queue.
-     * \return the element with the highest priority value
-     */
+	/*!
+	 * Return a reference to the highest-priority element. This is an O(1) operation.
+	 * Keep in mind that the reference might be invalidated after modifying the queue.
+	 * \return the element with the highest priority value
+	 */
 	virtual const T& peek() const = 0;
 
 private:
@@ -57,12 +57,12 @@ private:
 
 template<typename T>
 priority_queue<T>::priority_queue()
-    : sort_comparator(standard_comparator<T>)
+	: sort_comparator(standard_comparator<T>)
 {}
 
 template<typename T>
 priority_queue<T>::priority_queue(comparator<T>& sort_comparator)
-    : sort_comparator(sort_comparator)
+	: sort_comparator(sort_comparator)
 {}
 
 #endif /* PRIORITY_QUEUES_PRIORITY_QUEUE_H_ */
