+++ src/priority_queues/priority_queue_array.h
@@ -19,31 +19,31 @@ class priority_queue_array : public priority_queue<T>
 	sorted_vector<T> sorted_data;
 
 public:
-    /*!
-     * \copydoc priority_queue::priority_queue()
-     */
-    priority_queue_array() = default;
-    /*!
-     * \copydoc priority_queue::priority_queue(comparator<T>&)
-     */
+	/*!
+	 * \copydoc priority_queue::priority_queue()
+	 */
+	priority_queue_array() = default;
+	/*!
+	 * \copydoc priority_queue::priority_queue(comparator<T>&)
+	 */
 	explicit priority_queue_array(comparator<T>& comparator);
 
 	~priority_queue_array() = default;
-    /*!
-     * \copydoc priority_queue::add(const T& element)
-     * \details The array-based implementation of this function works in O(n) time,
-     * possibly resizing the array and invalidating pointers.
-     */
+	/*!
+	 * \copydoc priority_queue::add(const T& element)
+	 * \details The array-based implementation of this function works in O(n) time,
+	 * possibly resizing the array and invalidating pointers.
+	 */
 	void add(const T& element) override;
-    /*!
-     * \copydoc priority_queue::pop()
-     * \details The array-based implementation of this function works in O(1) time.
-     */
+	/*!
+	 * \copydoc priority_queue::pop()
+	 * \details The array-based implementation of this function works in O(1) time.
+	 */
 	T pop() override;
-    /*!
-     * \copydoc priority_queue::peek()
-     */
-    const T& peek() const override;
+	/*!
+	 * \copydoc priority_queue::peek()
+	 */
+	const T& peek() const override;
 
 private:
 	void print(std::ostream& os) const override;
