+++ src/priority_queues/priority_queue_list.h
@@ -29,29 +29,29 @@ private:
 	list_node<T>* root = nullptr;
 
 public:
-    /*!
-     * \copydoc priority_queue::priority_queue()
-     */
-    priority_queue_list() = default;
-    /*!
-     * \copydoc priority_queue::priority_queue(comparator<T>&)
-     */
-    explicit priority_queue_list(comparator<T>& comparator);
+	/*!
+	 * \copydoc priority_queue::priority_queue()
+	 */
+	priority_queue_list() = default;
+	/*!
+	 * \copydoc priority_queue::priority_queue(comparator<T>&)
+	 */
+	explicit priority_queue_list(comparator<T>& comparator);
 	~priority_queue_list();
 
-    /*!
-     * \copydoc priority_queue::add(const T& element)
-     * \details The linked-list-based implementation of this function operates in O(n) time.
-     */
+	/*!
+	 * \copydoc priority_queue::add(const T& element)
+	 * \details The linked-list-based implementation of this function operates in O(n) time.
+	 */
 	void add(const T& element) override;
-    /*!
-     * \copydoc priority_queue:pop()
-     * \details The linked-list-based implementation of this function operates in O(1) time.
-     */
+	/*!
+	 * \copydoc priority_queue:pop()
+	 * \details The linked-list-based implementation of this function operates in O(1) time.
+	 */
 	T pop() override;
-    /*!
-     * \copydoc priority_queue::peek()
-     */
+	/*!
+	 * \copydoc priority_queue::peek()
+	 */
 	const T& peek() const override;
 
 private:
