+++ src/vector.h
